+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 --------------
 
 - **Fix**: Fix diagnostic when an `@Multibinds`-annotated Map uses an enum as a key.
+- **Fix**: Fix Dagger interop error messages when validating `@Module`s with constructor-injected fields to match errors for `@BindingContainer`s.
 
 0.7.5
 -----